% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_judgement.R
\name{ahp.interval}
\alias{ahp.interval}
\title{A function to obtain the probabilities of rank reversals}
\usage{
ahp.interval(lower, upper, nmatrices, norm_test)
}
\arguments{
\item{lower}{A list of dataframes objects with the lower bounds for each pairwise comparison matrix}

\item{upper}{A list of dataframes objects with the upper bounds for each pairwise comparison matrix}

\item{nmatrices}{Integer, defines the total number of random matrices to generate}

\item{norm_test}{Logical, if TRUE determines if the components of the right eigenvector of all the generated random
pairwise comparison matrices are normally distributed, via the kolmogorov-smirnoff test}
}
\value{
A list of list, for each interval pairwise comparison matrix
\describe{
  \item{matrix_list}{List of random matrices (rm)}
  \item{matrices_w}{List of weight of rm}
  \item{norm_matrices_w}{List of normalized weights of rm}
  \item{lambda_max}{List of largest eigenvalues}
  \item{ci}{List of consistency indices}
  \item{cr}{List of consistency ratios}
  \item{w_consistent}{Dataframe of normalized weights of consistent random matrices (crm)}
  \item{maximum}{Named vector, maximum value of the crm for each criteria}
  \item{minimum}{Named vector, minimum value of the crm for each criteria}
  \item{mean}{Named vector, mean value of the crm for each criteria}
  \item{sd}{Named vector, standard deviation value of the crm for each criteria}
  \item{normality}{List of list with the normality test for each criteria}
  \item{p_ij}{Vector, probabilities of rank reversal between two alternatives \eqn{A_i} and \eqn{A_j}}
  \item{p_i}{Vector, probabilities that a given alternative will reverse rank with other alternative}
}
}
\description{
A function to obtain the probabilities of rank reversals
}
\section{Details}{

The probabilities \eqn{p_i}, \eqn{i = 1, \dots , n} that a given alternative will reverse
rank with another alternative are given by \deqn{p_i=1-\prod_{j=1}^n (1-p_{ij})} where
\eqn{p_{ij}} is the probability of rank reversal for two alternatives, it is calculated considering
different cases for the lower and upper bounds on the \eqn{i} and \eqn{j} components of the right eigenvector \eqn{w} and the
probability cumulative distribution \eqn{F_i(x_i)}:
\tabular{lll}{
Case \tab Condition \tab \eqn{p_{ij}} \cr
1 \tab \eqn{w_j^L\leq w_i^L} and \eqn{w_i^U\leq w_j^U} \tab \eqn{F_j(w_i^U)-F_j(w_i^L)} \cr
2 \tab \eqn{w_i^L < w_j^L} and \eqn{w_j^U < w_i^U} \tab \eqn{F_i(w_j^U)-F_i(w_j^U)} \cr
3 \tab \eqn{w_i^L < w_j^L < w_i^U < w_j^U} \tab \eqn{(F_i(w_i^U)-F_i(w_j^L))(F_j(w_i^U)-F_j(w_j^L))}\cr
4 \tab \eqn{w_j^L < w_i^L < w_j^U < w_i^U} \tab \eqn{(F_i(w_j^U)-F_i(w_i^L))(F_j(w_j^U)-F_j(w_i^L))} \cr
}
}

\examples{

}
\references{
Saaty, Thomas L. y Vargas, Luis G.: Uncertainty and rank order in the analytic hierarchy
process. European Journal of Operational Research, 1987, 32(1), pp. 107–117. ISSN 0377-
 2217. doi: 10.1016/0377-2217(87)90275-X.
}
\keyword{comparison}
\keyword{interval}
\keyword{matrix}
\keyword{pairwise}
\keyword{rank}
\keyword{reversal,}
