% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahpsensitivity-package.R
\docType{package}
\name{ahpsensitivity-package}
\alias{ahpsensitivity}
\alias{ahpsensitivity-package}
\title{ahpsensitivity: Sensitivity analysis tools for Multicriteria Decision Making methods}
\description{
Contains a set of sensitivity analysis tools for two Multicriteria Decision 
	Making (MDCM) methods: Analytic Hierarchy Process (AHP) and Weighted Sum 
	Model (WSM). For the AHP, the probability of rank reversals and the most 
	influential user can be obtained. Moreover, for both methods, the most critical measure of performance can be determined.
}
\author{
\strong{Maintainer}: Ileana Grave \email{igrave@iecologia.unam.mx} (\href{https://orcid.org/0000-0002-1389-1768}{ORCID})

Authors:
\itemize{
  \item Luis Bojorquez-Tapia \email{luis@multicriteria.org}
  \item Alejandra Estrada-Baron \email{ale.esba@gmail.com}
}

}
\keyword{internal}
