% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_judgement.R
\name{consistency.index}
\alias{consistency.index}
\title{A function to obtain the consistency index}
\usage{
consistency.index(lambda_max, n)
}
\arguments{
\item{lambda_max}{largest eigenvalue}

\item{n}{total number of criteria}
}
\value{

}
\description{
A function to obtain the consistency index
}
\section{Details}{

\deqn{C.I. = \frac{\lambda_{max}-n}{n-1}} where
\eqn{\lambda_{max}} is the largest eigenvalue and \eqn{n} represents the total number of
criteria
}

\examples{
consistency.index(lambda_max = 1, n = 6)
}
\keyword{consistency}
\keyword{eigenvalue}
\keyword{index,}
\keyword{largest}
