% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_judgement.R
\name{consistency.ratio}
\alias{consistency.ratio}
\title{A function to obtain the consistency ratio}
\usage{
consistency.ratio(ci, ri)
}
\arguments{
\item{ci}{Consistency index}

\item{ri}{Random inconsistency}
}
\value{

}
\description{
A function to obtain the consistency ratio
}
\note{
The random inconsistency number is a function of the number of criteria
}
\section{Details}{

\deqn{C.R. = \frac{C.I.}{R.I.}}
}

\examples{
consistency.ratio(.05, 0.90)
}
