% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_sensitivity.R
\name{group.sens}
\alias{group.sens}
\title{Group sensitivity}
\usage{
group.sens(users_list, pert)
}
\arguments{
\item{users_list}{List with the pairwise comparison matrices}

\item{pert}{Double, the perturbation magnitude for the finite differences method}
}
\value{
A list of lists for each user
\describe{
     \item{users_list}{List with the pairwise comparison matrices}
     \item{n}{Number of criteria}
     \item{users_total}{Integer, the total number of users}
     \item{pert}{Double, the perturbation magnitude for the finite differences method}
     \item{users_w}{List with weights for the pairwise comparison matrices}
     \item{norm_users_w}{List with the normalized weights for the pairwise comparison matrices}
     \item{group_matrix}{Matrix with the normalized weight for all criteria (rows: criteria, columns: users)}
     \item{sensitivity}{List of matrices with the sensitivity coefficients for each user}
}
}
\description{
This function allows you to identify the most influential expert on the group
}
\references{
Marie Ivanco, Gene Hou, Jennifer Michaeli, Sensitivity analysis method to address user disparities in the analytic hierarchy process,
Expert Systems with Applications, Volume 90, 2017, Pages 111-126, ISSN 0957-4174,
https://doi.org/10.1016/j.eswa.2017.08.003.
}
\keyword{ahp,}
\keyword{group}
\keyword{sensitivity,}
