% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_sensitivity.R
\name{group.vector}
\alias{group.vector}
\title{Create a group vector}
\usage{
group.vector(x, users_total)
}
\arguments{
\item{x}{a list of matrices with the normalized weight for all criteria and users (rows: critera, columns: user)}

\item{users_total}{the total number of users in the group}
}
\value{
A list of vectors
}
\description{
This function allows you to create the group vector
}
\section{Details}{
 The group aggregated vector is obtained by following
\deqn{{G}_r^{(i, j_{i-1})}= \sqrt[P]{\bar{w}_r^1\bar{w}_r^2\dots \bar{w}_r^P}}
where \eqn{P} represents the total number of users in the group and
\eqn{\bar{w}_r} is the normalized weight of the rth criterion
}

\references{
Marie Ivanco, Gene Hou, Jennifer Michaeli, Sensitivity analysis method to address user disparities in the analytic hierarchy process,
Expert Systems with Applications, Volume 90, 2017, Pages 111-126, ISSN 0957-4174,
https://doi.org/10.1016/j.eswa.2017.08.003.
}
\keyword{group}
\keyword{vector}
