% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_sensitivity.R
\name{norm.fun}
\alias{norm.fun}
\title{The normalized criteria weight}
\usage{
norm.fun(x)
}
\arguments{
\item{x}{Vector of criteria weight}
}
\value{
A vector
}
\description{
This function allows you to obtain the normalized criteria weight
(normalized geometric mean) for one user
}
\section{Details}{
 For each component \eqn{r} of the vector the normalized weight
is obtained by
\deqn{\bar{w}_r = \frac{w_r}{\sum_{q=1}^n w_q}}
where \eqn{n} is the total number of criteria
}

\examples{
w <- c(.7,.2,.3)
norm.fun(w)
}
\keyword{normalized}
\keyword{weights}
