% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_judgement.R
\name{pcmatrix}
\alias{pcmatrix}
\title{A function to create a pairwise comparison matrix}
\usage{
pcmatrix(x, n)
}
\arguments{
\item{x}{Vector, it contains the upper triangular elements of the matrix}

\item{n}{Integer, total number of criteria}
}
\value{

}
\description{
A function to create a pairwise comparison matrix
}
\section{Details}{

Given a vector with the upper triangular elements of a pairwise comparison matrix,
this function returns a matrix in the pairwise comparison form
\deqn{C = \left[ {\begin{array}{cccc}
 1 & a_{12} & \dots & a_{1m}\\
 1/a_{12} & 1 & \dots & a_{2m} \\
 \dots & \dots & 1 & \dots\\
 1/a_{1m} & 1/a_{2m} & \dots &  1
 \end{array} } \right]}
}

\examples{
pcmatrix(c(1,2,3,4,5,6), 4)
}
\keyword{comparison}
\keyword{matrix}
\keyword{pairwise}
