% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_judgement.R
\name{random.pcmatrices}
\alias{random.pcmatrices}
\title{A function to obtain random matrices}
\usage{
random.pcmatrices(data_lower, data_upper, nmatrices)
}
\arguments{
\item{data_lower}{A list of dataframes objects with the lower bounds for each pairwise comparison
matrix}

\item{data_upper}{A list of dataframes objects with the upper bounds for each pairwise comparison
matrix}

\item{nmatrices}{Integer, total number of random matrices to generate}
}
\value{
A list of matrices
}
\description{
A function to obtain random matrices
}
\section{Details}{

Given matrices
\deqn{C^{L} = \left[ {\begin{array}{cccc}
 1 & a^L_{12} & \dots & a^L_{1m}\\
 1/a^L_{12} & 1 & \dots & a^L_{2m} \\
 \dots & \dots & 1 & \dots\\
 1/a^L_{1m} & 1/a^L_{2m} & \dots &  1
 \end{array} } \right],\hspace{.5cm}
 C^{U} = \left[ {\begin{array}{cccc}
 1 & a^U_{12} & \dots & a^U_{1m}\\
 1/a^U_{12} & 1 & \dots & a^U_{2m} \\
 \dots & \dots & 1 & \dots\\
 1/a^U_{1m} & 1/a^U_{2m} & \dots &  1
 \end{array} } \right]}
with the lower and upper bounds of an interval pairwise comparison
matrix, this function generates a set of random matrices within the interval. The random
numbers for each comparison are generated with an uniform distribution.
}

\references{
Saaty, Thomas L. y Vargas, Luis G.: Uncertainty and rank order in the analytic hierarchy
process. European Journal of Operational Research, 1987, 32(1), pp. 107–117. ISSN 0377-
 2217. doi: 10.1016/0377-2217(87)90275-X.
}
\seealso{
runif about the uniform distribution
}
\keyword{matrix}
\keyword{numbers,}
\keyword{random}
