% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_sensitivity.R
\name{weights.matrix}
\alias{weights.matrix}
\title{A function to obtain the criteria weight}
\usage{
\method{weights}{matrix}(x, n)
}
\arguments{
\item{x}{a pairwise comparison matrix}

\item{n}{Integer, the total number of criteria}
}
\description{
Given a pairwise comparison matrix this function allows you to obtain the
criteria  weight (row geometric mean)
for one user and any number of criteria
}
\section{Details}{

The weight criteria is obtained by the computation of the geometric mean
\deqn{\bar{w}_r^{P_j} = \sqrt[m]{a_{r1}a_{r2}\dots a_{rm}}}
where \eqn{a_{rm}} are elements in the pairwise comparisons matrix
\deqn{C = \left[ {\begin{array}{cccc}
 1 & a_{12} & \dots & a_{1m}\\
 1/a_{12} & 1 & \dots & a_{2m} \\
 \dots & \dots & 1 & \dots\\
 1/a_{1m} & 1/a_{2m} & \dots &  1
 \end{array} } \right]}
}

\examples{
pcm <- pcmatrix(c(1,2,3,4,5,6), 4)
weights.matrix(pcm, 4)
}
\references{
Marie Ivanco, Gene Hou, Jennifer Michaeli, Sensitivity analysis method to address user disparities in the analytic hierarchy process,
Expert Systems with Applications, Volume 90, 2017, Pages 111-126, ISSN 0957-4174,
https://doi.org/10.1016/j.eswa.2017.08.003.
}
\keyword{matrix}
\keyword{weights}
