% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{wsm.tau}
\alias{wsm.tau}
\title{A function to obtain the most critical measure of performance in terms of some
reference value for a set of indices}
\usage{
wsm.tau(x_ij, w_ij, indices, rho)
}
\arguments{
\item{x_ij}{A list of matrices with the standarized measures of performance
for each index \eqn{i} and each criteria \eqn{j}}

\item{w_ij}{A list of matrices with the importance weights for each criteria \eqn{j}}

\item{indices}{A list of dataframes with the preferences for each alternative}

\item{rho}{A list with the references values for each index}
}
\value{



}
\description{
A function to obtain the most critical measure of performance in terms of some
reference value for a set of indices
}
\section{Details}{

For each index \eqn{i}, the minimum change in a measure of performance \eqn{x^h_{ij}} for a criteria \eqn{j} that
generates a rank reversal is obtained with
\deqn{\tau_{ij}^h = \frac{V_i^{h}-V_i^{\rho}}{w_{ij}x^h_{ij}}}
In terms of \eqn{\tau_{ij}^h}, the critical indicator value
\deqn{\mathcal{C}_{ij} = \frac{1}{\Delta_{ij}}\times p_{ij}}
considers the sensitivity coefficient \eqn{\Delta_{ij}=|\tau_{ij}^h|^{Q_1}}
(first quartile \eqn{Q_1}) and the probability of rank reversals \eqn{p_{ij}}.
Moreover, the modified measure of performance is given by
\deqn{\hat{x}_{ij}^h = {x}_{ij}^h - \tau_{ij}^h}
}

\references{
Sensitivity analysis for household vulnerability assessment: a case of study from Brazil surveys

Triantaphyllou, Evangelos y Sánchez, Alfonso: A Sensitivity Analysis Approach for Some Deterministic
Multi-Criteria Decision-Making Methods*. Decision Sciences, 1997, 28(1), pp. 151–194.
doi: 10.1111/j.1540-5915.1997.tb01306.x.
https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1540-5915.1997.tb01306.x
}
